+++ plone/app/contenttypes/tests/test_behaviors_table_of_contents.py
@@ -26,7 +26,7 @@ class DocumentFunctionalTest(unittest.TestCase):
         self.portal = self.layer['portal']
         self.request = self.layer['request']
         self.portal_url = self.portal.absolute_url()
-        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        setRoles(self.portal, TEST_USER_ID, ['Contributor'])
         fti = DexterityFTI('tocdocument')
         self.portal.portal_types._setObject('tocdocument', fti)
         fti.klass = 'plone.dexterity.content.Item'
@@ -68,7 +68,3 @@ class DocumentFunctionalTest(unittest.TestCase):
         # Submit form
         self.browser.getControl('Save').click()
         self.assertTrue('<dl id="document-toc"' in self.browser.contents)
-
-
-def test_suite():
-    return unittest.defaultTestLoader.loadTestsFromName(__name__)